Use 3.3V power input
The OpenLog features a voltage regulator that steps input power down to 3.3V.
According to the ATmega328P data sheet (section 28, Electrical Characteristics),
the maximum voltage on any pin is VCC + 0.5 = 3.8V.

The SP3232E transceiver's voltage level on RxOUT = VCC + 0.3. Therefore the
transceiver's VCC cannot exceed 3.5V or it will be out of spec with the ATmega.